services/bot/requirements.txt
psycopg2-binary
pydantic-settings
sqlalchemy
httpx